requirements.txt
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.1
dj-database-url==1.2.0
Django==4.1.4
django-configurations==2.4
django-cors-headers==3.13.0
djangorestframework==3.14.0
drf-yasg==1.21.4
executing==1.2.0
idna==3.4
inflection==0.5.1
ipdb==0.13.11
ipython==8.8.0
itypes==1.2.0
jedi==0.18.2
Jinja2==3.1.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mercadopago==2.2.0
packaging==22.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
prompt-toolkit==3.0.36
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.14.0
pytz==2022.7
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
tomli==2.0.1
traitlets==5.8.1
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.13
wcwidth==0.2.5
psycopg2-binary==2.9.3 
gunicorn==20.1.0
